refacto: DTMT-11 Remove IdentifierSorter
