--- src/main/kotlin/datamaintain/sort/AlphabeticalSorter.kt
@@ -1,10 +1,0 @@
-package datamaintain.sort
-
-import datamaintain.Config
-import datamaintain.Script
-
-class AlphabeticalSorter(config: Config) : Sorter(config) {
-    override fun <T : Script> sort(scripts: List<T>): List<T> {
-        return scripts.sortedBy {script -> script.name.toLowerCase()}
-    }
-}
