+++ src/main/kotlin/datamaintain/sort/Sorter.kt
@@ -3,6 +3,6 @@ package datamaintain.sort
 import datamaintain.Config
 import datamaintain.Script
 
-abstract class Sorter(protected val config: Config) {
-    abstract fun <T : Script> sort(scripts: List<T>): List<T>
+abstract class Sorter<U>(protected val config: Config) {
+    abstract fun <T : Script> sort(scripts: List<T>, getter: (T) -> U): List<T>
 }
